move `Impl::Handler` to a separate file
